Editting model to exclude bcrypt gem.
For this iteration, we need to not use any gems, including the bcrypt gem. With that in mind, for this iteration, we won't have encrypted passwords. Thus, in order to authenticate users, we'll need to simply just check if the inputted password matches the password for that user in the db.